requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
google-generativeai==0.7.2
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.8.2
requests==2.32.3
